Update config reading for local, default, and optional
